requirements.txt
torch
transformers
Pillow
numpy
opencv-python
# For CLIP:
ftfy
regex
tqdm
wandb
git+https://github.com/openai/CLIP.git
fastapi
python-multipart
jinja2
uvicorn